Break video into smaller files
